Remove material design lite.
Bringing in MDL causes conflicts with youtube's stylesheet. The only
effect visible was that the video control buttons were misaligned, but
this pointed to a larger problem that MDL's CSS could unknowingly
break youtube's CSS anywhere on the site. I attempted to try to use
the shadow DOM to encapsulate the CSS, but there doesn't seem to be
reliable tooling around injecting CSS into the shadow DOM. It's much
easier to simply write my own styles and scope them to my own widgets.